updated txt output files after last commit and updated some test files

+++ src/tests/testfiles/programs/areaCirclePointer.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 void areaperi ( int r, float *a, float *p ) {
-    *a = 3.14 * r * r;
-    *p = 2 * 3.14 * r;
+    *a = 3.14 * (float) r * (float) r;
+    *p = (float) 2 * 3.14 * (float) r;
 }
 
 void main( ) {
