+++ src/tests/testfiles/programs/centigradeToFarenheit.c
@@ -6,7 +6,7 @@ int main() {
 	printf("\nEnter temp in Celsius : ");
 	scanf("%f", &celsius);
 
-	fahrenheit = (1.8 * celsius) + 32;
+	fahrenheit = (1.8 * celsius) + (float) 32;
 	printf("\nTemperature in Fahrenheit : %f ", fahrenheit);
 
 	return (0);
